requirements.txt
streamlit>=1.28.0
streamlit-authenticator>=0.2.3
bcrypt>=4.0.1
pyyaml>=6.0

